+++ src/lib/broadcast-to-child-windows.ts
@@ -13,7 +13,9 @@ export function broadcastToChildWindows(
     if (childWindow.closed) {
       childWindows.splice(i, 1);
     } else if (source !== childWindow) {
-      broadcast(childWindow.top, payload, origin, limitBroadCastToOrigin);
+      if (childWindow.top) {
+        broadcast(childWindow.top, payload, origin, limitBroadCastToOrigin);
+      }
     }
   }
 }