+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 5.2.0
+
+- Add support for `limitBroadCastToOrigin` as a framebus config option
+
 # 5.1.2
 
 - Fix issue where framebus could not be used with server side rendering